Updated Many Container Versions
Updated Rust Dockerfile, and cargo dependency versions. Removed Haskell
subscriber as that is not really pragmatic for what we want to
accomplish here. Updated the publisher to start sending out jobs to have
subscribers do that will do work and then generate metrics. I am running
into libssl issues that persist despite the run image in the
Dockerfiles, which I believe is related to an outdated WSL image.
Upgrading that is not feasable at this time, so I am saving this work to
continue to test on a more updated system.